feat(components): add deployed bytecode
