+++ src/lib/pages/interact-contract/components/index.ts
@@ -1,4 +1,3 @@
 export * from "./execute-area";
 export * from "./query-area";
-export * from "./InteractionTypeSwitch";
 export * from "./InteractionWrapper";
